// src/types/enum.ts
export enum BasicStatus {
  DISABLE = 0,
  ENABLE = 1
}

export enum ResultEnum {
  SUCCESS = 0,
  ERROR = -1,
  TIMEOUT = 401
}

export enum StorageEnum {
  UserInfo = 'userInfo',
  AccessToken = 'accessToken',
  RefreshToken = 'refreshToken',
  Settings = 'settings',
  I18N = 'i18nextLng'
}

export enum ThemeMode {
  Light = 'light',
  Dark = 'dark'
}

export enum ThemeLayout {
  Vertical = 'vertical',
  Horizontal = 'horizontal',
  Mini = 'mini'
}

export enum ThemeColorPresets {
  Default = 'default',
  Cyan = 'cyan',
  Purple = 'purple',
  Blue = 'blue',
  Orange = 'orange',
  Red = 'red'
}

export enum LocalEnum {
  en_US = 'en_US',
  zh_CN = 'zh_CN'
}

export enum MultiTabOperation {
  FULLSCREEN = 'fullscreen',
  REFRESH = 'refresh',
  CLOSE = 'close',
  CLOSEOTHERS = 'closeOthers',
  CLOSEALL = 'closeAll',
  CLOSELEFT = 'closeLeft',
  CLOSERIGHT = 'closeRight'
}

export enum PermissionType {
  CATALOGUE = 0,
  MENU = 1,
  BUTTON = 2
}

export enum HtmlDataAttribute {
  ColorPalette = 'data-color-palette',
  ThemeMode = 'data-theme-mode'
}